Implemented Employee class.
I believe I have finished implementation on this as per the documentation, letr me know if you run into any issues with it.

I went ahead and had the employee class add the created event to the database, but if this makes more sense elsewhere, Feel free to move it.

Finally, I did run into some GUI issues with things not seeming to be fitted to my screen and the attendees list not showing on the create event screen. I am not sure exaclty how to fix this as of current.