More bugfixes & added time conflict check
Added check for time conflict in events.
With this I fixed some issues related to 12:00 AM being considered after 1:00 AM by the program because 12 > 1, now uses military time internally which doesn't affect anything else so it's no cause for concern.
With this I made some changes to the delete event function as this was also affected somewhat by these issues.