Even More Bug Fixes
The event list box was causing so many issues so I looked further into it and replaced it with an event view list that works 1000 times better. I also worked to make sure that the eventID is provided to the program so that way we can edit events by the ID rather than their values.